svelte/test-svelte-app/backend/requirements_gpu.txt
fastapi
pydantic
python-multipart
uvicorn
transformers
soundfile
torchaudio
accelerate
nemo_toolkit[asr]>=1.20.0